software/docker/api/requirements.txt
Flask
retry
requests
uWSGI
redis
pillow